refactor the code to add content read & other requested features
