+++ lib/src/download_file_request.dart
