+++ lib/src/download_task.dart
