+++ lib/src/notification_types.dart
